examples/todo-app-with-oauth2/requirements.txt
requests
flask
sqlalchemy
psycopg2-binary
authlib
python-dotenv
werkzeug